backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
redis>=5.0.0
alembic>=1.13.0
httpx>=0.27.0
websockets>=12.0
certifi>=2024.0.0
anthropic>=0.39.0
weasyprint>=62.0
jinja2>=3.1.0
Pillow>=10.0.0
sentry-sdk[fastapi]>=1.39.0
